Minor update to simplify internals
You now pass in the list of choices as an array of strings and the component will take care of creating the special Dropdown option objects required for the Semantic UI dropdown component.  Also, if the default value to show in the dropdown is not in the options array passed in, that value will be automatically added so that it appears correctly.